requirements.txt
bcrypt
cryptography
mysql-connector-python
